Move labeling functions to util package
Labeling is needed not only in the template package but also elsewhere
